recruitmetsoft/requirements.txt
django==6.0.2
djangorestframework
djangorestframework-simplejwt
django-cors-headers
